# Conflicts:
#	src/main/java/com/yunche/loan/service/impl/BankSolutionServiceImpl.java
#	src/main/resources/sysconfig.properties

+++ src/main/java/com/yunche/loan/domain/param/CustomerParam.java
@@ -107,7 +107,7 @@ public class CustomerParam {
 
     private Long feedingNum;
 
-    private Long floorSpace;
+    private BigDecimal floorSpace;
 
     private Byte houseOwnerRelation;
